exec routes different from posix to windows
